+++ src/graphics/draw_calls.cpp
@@ -966,7 +966,7 @@ int32_t DrawCalls::getSkinningOffset() const
 {
     return std::accumulate(m_mesh_for_skinning.begin(),
         m_mesh_for_skinning.end(), 0, []
-        (const size_t previous, const STKAnimatedMesh* m)
+        (const unsigned int previous, const STKAnimatedMesh* m)
         { return previous + m->getTotalJoints(); });
 }   // getSkinningOffset
 #endif   // !SERVER_ONLY
