+++ src/graphics/post_processing.hpp
@@ -93,7 +93,7 @@ public:
 	void renderHorizontalBlur(const FrameBuffer &in_fbo, const FrameBuffer &auxiliary) const;
 
     void renderGaussian6BlurLayer(const FrameBuffer &in_fbo, const FrameBuffer &scalar_fbo,
-                                  size_t layer, float sigmaH, float sigmaV) const;
+                                  unsigned int layer, float sigmaH, float sigmaV) const;
     void renderGaussian17TapBlur(const FrameBuffer &in_fbo,
                                  const FrameBuffer &auxiliary,
                                  const FrameBuffer &linear_depth) const;
