+++ src/input/input_manager.cpp
@@ -1034,7 +1034,7 @@ EventPropagation InputManager::input(const SEvent& event)
     else if (event.EventType == EET_TOUCH_INPUT_EVENT)
     {
         MultitouchDevice* device = m_device_manager->getMultitouchDevice();
-        unsigned int id = event.TouchInput.ID;
+        unsigned int id = (unsigned int)event.TouchInput.ID;
 
         if (device != NULL && id < device->m_events.size())
         {
