+++ src/states_screens/race_gui_multitouch.cpp
@@ -271,8 +271,8 @@ void RaceGUIMultitouch::drawMultitouchSteering(const AbstractKart* kart,
                 float scale = UserConfigParams::m_multitouch_scale *
                     (float)(irr_driver->getActualScreenSize().Height) / 720.0f;
 
-                m_race_gui->drawEnergyMeter(button->x + button->width * 1.15f,
-                                            button->y + button->height * 1.35f,
+                m_race_gui->drawEnergyMeter(int(button->x + button->width * 1.15f),
+                                            int(button->y + button->height * 1.35f),
                                             kart, viewport,
                                             core::vector2df(scale, scale));
             }
